nlu/requirements.txt
rasa_nlu==0.14.4
spacy==2.0.18
gevent==1.2.2
klein==17.10.0
hyperlink==17.3.1
typing==3.6.2
future==0.17.1
six==1.11.0
jsonschema==2.6.0
matplotlib==2.1.0
requests==2.20.0
tqdm==4.19.5
numpy==1.14.5
simplejson==3.13.2
cloudpickle==0.6.1
msgpack-python==0.5.4
packaging==18.0
pyyaml==3.12
coloredlogs==10.0
ruamel.yaml==0.15.78
google-cloud-storage
fuzzy-matcher
requests-futures==0.9.7
python-Levenshtein
tensorflow==1.12.0
scipy==1.1.0
sklearn-crfsuite==0.3.6
scikit-learn==0.20.2